Rename constants to defaults
